demo_project/requirements.txt
azure-common==1.1.18
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b4
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==1.5.0
azure-storage-common==1.4.0
